Splitting code and cleaning up (#13)
* Splitting modules

* Splitting env vars

* Adding import os

* config change

* correct spelling

* changing register route
 to user

* removing debug log

* fetch tokens after reject

* cleanup

* cleanup
